refactor: move svg assets to 'assets/icons' folder
